refactor: move the path utility function file to the /src folder and fix all related files using that file
