+++ packages/untp-test-suite/src/templates/mapper.ts
@@ -3,7 +3,7 @@ import fs from 'fs/promises';
 import path from 'path';
 import _ from 'lodash';
 import { TestSuiteResult } from '../core/types/index.js';
-import { getCurrentDirPath, getCurrentFilePath } from './utils/path.js';
+import { getCurrentDirPath, getCurrentFilePath } from '../utils/path.js';
 
 Handlebars.registerHelper('jsonStringify', (jsonObject: object) => {
   if (!_.isObject(jsonObject) || _.isEmpty(jsonObject)) {
@@ -18,7 +18,7 @@ Handlebars.registerHelper('jsonStringify', (jsonObject: object) => {
 export async function templateMapper(templateName: string, testSuiteResult: TestSuiteResult) {
   try {
     const currentDirPath = getCurrentDirPath(getCurrentFilePath());
-    const templateFilePath = path.join(currentDirPath, `../templateMessages/${templateName}.hbs`);
+    const templateFilePath = path.join(currentDirPath, `../templates/templateMessages/${templateName}.hbs`);
     const template = await fs.readFile(templateFilePath, 'utf-8');
 
     const compiledTemplate = Handlebars.compile(template);
